+++ app/src/main/java/org/simple/clinic/search/results/PatientSearchResultsScreen.kt
@@ -15,7 +15,6 @@ import org.simple.clinic.facility.Facility
 import org.simple.clinic.facility.alertchange.AlertFacilityChangeSheet
 import org.simple.clinic.facility.alertchange.Continuation.ContinueToScreen
 import org.simple.clinic.navigation.v2.Router
-import org.simple.clinic.navigation.v2.compat.wrap
 import org.simple.clinic.navigation.v2.fragments.BaseScreen
 import org.simple.clinic.newentry.PatientEntryScreenKey
 import org.simple.clinic.patient.PatientSearchCriteria
@@ -131,11 +130,11 @@ class PatientSearchResultsScreen :
   }
 
   override fun openPatientSummaryScreen(patientUuid: UUID) {
-    router.push(PatientSummaryScreenKey(patientUuid, OpenIntention.ViewExistingPatient, Instant.now(utcClock)).wrap())
+    router.push(PatientSummaryScreenKey(patientUuid, OpenIntention.ViewExistingPatient, Instant.now(utcClock)))
   }
 
   override fun openLinkIdWithPatientScreen(patientUuid: UUID, identifier: Identifier) {
-    router.push(PatientSummaryScreenKey(patientUuid, OpenIntention.LinkIdWithPatient(identifier), Instant.now(utcClock)).wrap())
+    router.push(PatientSummaryScreenKey(patientUuid, OpenIntention.LinkIdWithPatient(identifier), Instant.now(utcClock)))
   }
 
   override fun openPatientEntryScreen(facility: Facility) {